Removed remaining instances of cereal in header files.
